added a cast to remove a compiler warning
git-svn-id: http://svn.eol.ucar.edu/svn/eldora/trunk@1816 15f24ed5-0439-0410-88a4-bd1d0fc06cdd

+++ redrapids/RR314isr.cpp
@@ -82,7 +82,7 @@ void processDMAGroups(s_ChannelAdapter *pCA, int chan, RR314* pRR314) {
 
 			// update the last group index to indicated the next
 			// group that is available.
-			if (pRR314->lastGroup(chan) < pCA->DMA.GrpCnt[chan]) {
+			if (pRR314->lastGroup(chan) < (int)pCA->DMA.GrpCnt[chan]) {
 				pRR314->lastGroup(chan, pRR314->lastGroup(chan)+1);
 			} else {
 				// handle the wrap around
